# docs/gpt/openapi.yaml
...
openapi: 3.1.0
info:
  title: Fishbowl GPT Content Tools
  version: 4.0.0
  description: Curate, rate and publish articles, and attach generated images.
servers:
  - url: https://creepender.top/api/gpt
security:
  - BearerAuth: []
paths:
  /articles/{id}:
    get:
      operationId: getArticle
      summary: Read editable article by ID extracted from its link. Fetch all pages at
        the same revision before curation or image planning. Content slices
        preserve original text.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticlePage"
        default:
          description: 400 invalid input; 401/403 authentication/permission; 404 missing;
            409 revision/request conflict; 413 too large; 500/504
            server/timeout.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            maxLength: 36
        - name: offset
          in: query
          required: false
          schema: &a1
            type: integer
            minimum: 0
            default: 0
        - name: block_offset
          in: query
          required: false
          schema: *a1
  /articles:
    get:
      operationId: searchArticles
      summary: Search editable article titles and excerpts, including drafts. Returns
        20 results per page.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Search"
        default:
          description: 400 invalid input; 401/403 authentication/permission; 404 missing;
            409 revision/request conflict; 413 too large; 500/504
            server/timeout.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - all
              - draft
              - published
            default: all
        - name: offset
          in: query
          required: false
          schema: *a1
...
  /save-rated:
    post:
      operationId: saveRatedArticle
      summary: "Create: title, summary, blocks with access_level, rating_reason
        required. Draft by default; status=published publishes. Existing
        article: send ONLY article_ref, expected_revision, request_id, status to
        publish/unpublish."
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
        default:
          description: 400 invalid input; 401/403 authentication/permission; 404 missing;
            409 revision/request conflict; 413 too large; 500/504
            server/timeout.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveRated"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    ArticlePage:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
        article_id:
          type: string
          maxLength: 36
        revision:
          type: string
          maxLength: 64
        status: &a6
          type: string
          enum:
            - draft
            - published
          default: draft
        block_count:
          type: integer
        tags:
          type: array
          items:
            type: string
            maxLength: 80
          maxItems: 500
        visible_access_level: &a2
          type: integer
          minimum: 1
          maximum: 5
          description: 1=P public, 2=G general, 3=M member, 4=A adult, 5=R root. Assess
            each block.
        full_access_level: *a2
        cover_access_level: *a2
        article:
          type: object
          additionalProperties: false
          properties:
            title:
              type: string
              maxLength: 255
            excerpt:
              type:
                - string
                - "null"
            article_type:
              type: string
              maxLength: 20
            category_id:
              type:
                - string
                - "null"
            summary:
              type:
                - string
                - "null"
            metadata:
              $ref: "#/components/schemas/Metadata"
            rating_reason:
              type:
                - string
                - "null"
        blocks:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              id:
                type: string
                maxLength: 36
              type:
                type: string
                maxLength: 20
              access_level: *a2
              media_id:
                type:
                  - string
                  - "null"
              order:
                type: integer
              content_json:
                type: string
                maxLength: 8000
              content_offset:
                type: integer
              content_length:
                type: integer
          maxItems: 10
        next:
          type:
            - object
            - "null"
          properties:
            offset:
              type: integer
            block_offset:
              type: integer
        hint:
          type: string
          maxLength: 500
...
    Error:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
        error:
          type: string
          maxLength: 2000
...
    Metadata:
      type: object
      additionalProperties: false
      properties:
        source_url:
          type: string
          maxLength: 2000
        language:
          type: string
          maxLength: 35
        description:
          type: string
          maxLength: 2000
    RatedBlock:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          enum:
            - text
            - code
            - image
        content:
          type: string
          maxLength: 60000
        language:
          type: string
          maxLength: 80
        media_id:
          type: string
          maxLength: 36
        title:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        access_level: *a2
      required:
        - type
        - access_level
    Receipt:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
        article_id:
          type: string
          maxLength: 36
        revision:
          type: string
          maxLength: 64
        status: *a6
        block_count:
          type: integer
        tags:
          type: array
          items:
            type: string
            maxLength: 80
          maxItems: 500
        visible_access_level: *a2
        full_access_level: *a2
        cover_access_level: *a2
        added_block_ids:
          type: array
          items:
            type: string
            maxLength: 36
          maxItems: 5
    SaveRated:
      type: object
      additionalProperties: false
      properties:
        request_id: *a4
        article_ref: *a3
        expected_revision: *a5
        title:
          type: string
          maxLength: 255
        summary:
          type: string
          maxLength: 2000
        excerpt:
          type: string
          maxLength: 500
        tags:
          type: array
          items:
            type: string
            maxLength: 80
          maxItems: 10
        category_id:
          type: string
          maxLength: 36
        article_type:
          type: string
          enum:
            - default
            - text
            - image
            - code
            - diary
            - drawing
          default: text
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/RatedBlock"
          maxItems: 80
          minItems: 1
        status: *a6
        rating_reason:
          type: string
          maxLength: 2000
        metadata:
          $ref: "#/components/schemas/Metadata"
      required:
        - request_id
    Search:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
        articles:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              id:
                type: string
                maxLength: 36
              title:
                type: string
                maxLength: 255
              excerpt:
                type:
                  - string
                  - "null"
              status: *a6
              type:
                type: string
                maxLength: 20
              category_id:
                type:
                  - string
                  - "null"
          maxItems: 20
        next_offset:
          type:
            - integer
            - "null"